services/tweet-crawler/requirements.txt
attrs==19.3.0
blis==0.4.1
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
cymem==2.0.3
idna==2.9
importlib-metadata==1.5.0
kafka-python==2.0.1
more-itertools==8.2.0
murmurhash==1.0.2
numpy==1.18.1
oauthlib==3.1.0
packaging==20.3
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
py==1.8.1
pyparsing==2.4.6
PySocks==1.7.1
pytest==5.3.5
python-dateutil==2.8.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
spacy==2.2.3
srsly==1.0.2
thinc==7.3.1
tqdm==4.43.0
tweepy==3.8.0
urllib3==1.25.8
vaderSentiment==3.2.1
wasabi==0.6.0
wcwidth==0.1.8
zipp==3.1.0
